Move validate_registration() to ValidatorRegistrar
