+++ config.toml
@@ -1,0 +1,11 @@
+[boost]
+host = "0.0.0.0"
+port = 18550
+relays = [
+    "http://127.0.0.1:28545",
+]
+
+[relay]
+host = "0.0.0.0"
+port = 28545
+beacon_node_url = "http://192.168.1.12:5052"
